+++ src/main/java/roomescape/controller/api/ReservationController.java
@@ -1,21 +1,20 @@
-package roomescape.controller;
+package roomescape.controller.api;
 
 import jakarta.validation.Valid;
 import java.net.URI;
 import java.util.List;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 import roomescape.controller.request.CreateReservationRequest;
 import roomescape.controller.response.ReservationResponse;
 import roomescape.service.ReservationService;
 
-@Controller
+@RestController
 @RequestMapping("/reservations")
 public class ReservationController {
 
@@ -26,7 +25,7 @@ public class ReservationController {
     }
 
     @PostMapping
-    public ResponseEntity<ReservationResponse> reserve(@RequestBody @Valid final CreateReservationRequest request) {
+    public ResponseEntity<ReservationResponse> reserve(@Valid final CreateReservationRequest request) {
         var reservation = service.reserve(request.name(), request.date(), request.timeId(), request.themeId());
         var response = ReservationResponse.from(reservation);
         return ResponseEntity.created(URI.create("/reservations/" + reservation.id())).body(response);
