+++ src/main/java/roomescape/controller/api/ThemeController.java
@@ -1,23 +1,22 @@
-package roomescape.controller;
+package roomescape.controller.api;
 
 import jakarta.validation.Valid;
 import java.net.URI;
 import java.time.LocalDate;
 import java.util.List;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 import roomescape.controller.request.CreateThemeRequest;
 import roomescape.controller.response.ThemeResponse;
 import roomescape.service.ThemeService;
 
-@Controller
+@RestController
 @RequestMapping("/themes")
 public class ThemeController {
 
@@ -28,7 +27,7 @@ public class ThemeController {
     }
 
     @PostMapping
-    public ResponseEntity<ThemeResponse> register(@RequestBody @Valid final CreateThemeRequest request) {
+    public ResponseEntity<ThemeResponse> register(@Valid final CreateThemeRequest request) {
         var theme = service.register(request.name(), request.description(), request.thumbnail());
         var response = ThemeResponse.from(theme);
         return ResponseEntity.created(URI.create("/themes/" + response.id())).body(response);
